apps/livekit-agent/pyproject.toml
[project]
name = "sanne-livekit-agent"
version = "0.1.0"
description = "LiveKit voice worker for the Sanne demo"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[mistralai,silero,turn-detector]~=1.5",
    "livekit-plugins-noise-cancellation>=0.2.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
